+++ src/Server/Controllers/ClockController.cs
@@ -5,6 +5,7 @@ using Tellurian.Trains.MeetingApp.Clocks.Implementations;
 using Tellurian.Trains.MeetingApp.Contracts;
 using Tellurian.Trains.MeetingApp.Server.Extensions;
 using Tellurian.Trains.MeetingApp.Clocks;
+using System.Diagnostics;
 
 namespace Tellurian.Trains.MeetingApp.Server.Controllers
 {