+++ RELEASENOTES.md
@@ -19,6 +19,13 @@ In your brower settings, create an exception for clearing data on exit for the w
 This will keep your clock settings including your username and selected clock name.
 
 ## Releases
+### Release 3.4.0
+Release date 2023-07-14
+- **Pause reason** *Done for today* added. This is a breaking API change.
+- **Pause reasons changed**: *Coffe* changed to *CoffeBreak*, *Closing* changed to *HallIsClosing*. This is a breaking API change.
+- **Pause reasons** *Hall is closing* and *Done for today* now displays **Stops at ...** instead for **Pause at...**.
+- **Clock URL** now shows when clock has not started and no other message should be displayed.
+- - **Bug fix** of *Start* button not visible after pause when showing real time.
 ### Release 3.3.0
 Release date 2023-06-14
 - **Slovak** translation added. This is a preliminary translation that need to be checked and improved.